+++ client/package.json
@@ -10,6 +10,7 @@
         "@mui/x-charts": "^7.18.0",
         "@mui/x-data-grid": "^7.18.0",
         "@paypal/react-paypal-js": "^8.7.0",
+        "@react-pdf/renderer": "^4.0.0",
         "@reduxjs/toolkit": "^2.2.7",
         "@testing-library/jest-dom": "^5.17.0",
         "@testing-library/react": "^13.4.0",
